third_party/blink/web_tests/external/wpt/css/css-device-adapt/viewport-clamp-initial-scale-to-min.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Page with meta viewport "width=device-width, initial-scale=1.0, minimum-scale=2.0" should scale to 2.0.
  assert_equals: visual viewport scale should be 2.0 expected 2 but got 1
Harness: the test ran to completion.
